Damage lowers hitpoints, dead character is removed.

+++ imt3601_shineydonkeys/CombatComponent.h
@@ -18,7 +18,7 @@ public:
 
 	void startCombat(Character* other);
 	void endCombat();
-	void takeDamage(const float damage);
+	void takeDamage(const unsigned int damage);
 private:
 	bool _inCombat;
 	Character* _other;
